+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/HEAD)
 
-* 
+* Hide shards/mods if they are incompatible with the defense
 
 [Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.1.7...HEAD)